Fixed VERY annoying bug where the jobs were failng on ACCRE where it kept saying the .simg /OUTPUTS was read-only.
But the source of the problem was actually the input/output_dir_list.txt.
As reference, if you -B bind a garabage path to the OUTPUTS, you get the same error.
So it is something funky with the .txt file itself.

To fix I did the following:
I 'cat'd the .txt files on ACCRE.
Copied the list that was displayed.
renamed the files to OLD
'touch'd new files and copied the list in
Now it works... no idea what garbage was in the original .txt files to make it fail.